requirements.txt
python-dotenv
discord.py
requests
asyncio
espn_api
table2ascii